Updating User Class methods
Fixed error in addProduct method where a line could be skipped accidentally if a product already exists. 
Added writeUser() method, which will write the current user object to the userInfo.txt file. With this, the user class should have complete persistance as long as the methods are used correctly when dealing with user objects. *Make sure to use writeUser to save user info, and use the constructor with only the username constructor to load an already existing user!*
Also fixed toString() method so it creates a string that matches the format of lines in the userInfo.txt file.